requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
future==1.0.0
idna==3.7
install==1.3.5
pathvalidate==3.2.0
pydantic==2.7.0
pydantic_core==2.18.1
regex==2024.4.16
requests==2.31.0
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
